// libs/shared/header/src/lib/interfaces/apps.interface.ts
/*
 * Copyright © 2005 - 2021 Alfresco Software, Ltd. All rights reserved.
 *
 * License rights for this program may be obtained from Alfresco Software, Ltd.
 * pursuant to a written agreement and any use of this program without such an
 * agreement is prohibited.
 */

export interface UserAppsResponse {
    appKey: string;
    id: string;
    status: string;
    launchUrl: string;
    app: {
        id: string;
        localizedName: string;
        appType: string;
    };
}

export interface UserApps {
    appKey: string;
    launchUrl: string;
    localizedName: string;
}